% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebm_fit.R
\name{ebm_fit}
\alias{ebm_fit}
\title{Main function}
\usage{
ebm_fit(obs, forc, n_boxes, start_year = NULL, end_year = NULL,
  detrending = FALSE, config_file = "ebm_fit_config.yml",
  config = "experimental")
}
\arguments{
\item{obs}{Vector of temperature observation values or data frame with columns \code{year} and \code{temperature}.}

\item{forc}{Vector of forcing values or data frame with columns \code{year} and \code{forcing}.}

\item{n_boxes}{Number of boxes in the EBM.}

\item{start_year}{First of forcing considered (only relevant if \code{forc} or \code{obs} is a data frame,
otherwise entire vector is taken).}

\item{end_year}{Last year of forcing considered (inclusive; only relevant if \code{forc} or \code{obs} is a data frame,
otherwise entire vector is taken).}

\item{detrending}{Boolean - indicates if observations should be detrended linearly before applying the Bayesian fit.}

\item{config_file}{Path to config file or full config list.}

\item{config}{Name of config to use.}
}
\value{
Object of class \code{ebm_fit_result}, inherits from \code{mcmc_result}. In addition contains an entry
\itemize{
\item \code{meta}: list with entries
\itemize{
\item \code{n_boxes}
\item \code{start_year}
\item \code{end_year}
}
}

Just like objects of class \code{mcmc_result}, it also contains entries:
\itemize{
\item \code{input_params}: Input parameters (copy of \code{params})
\item \code{posteriors}: List that contains posterior means, medians and variance for each estimated parameter
\item \code{samples}: List with entries
\itemize{
\item \code{parameters}
(matrix of samples,  one row for each sample, one column for each variable)
\item \code{model_fit} (matrix of EBM solutions (for all samples (potentially large!),
one row for each sample, one column for each point in time).
Note the ordering: first the burn-in samples from all chains, then the rest.
}
\item \code{diagnostics}: (only if diagnostics is TRUE) list with diagnostic criteria
}
}
\description{
Fit EBM to data
}
