% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-MCMC.R
\name{extract_samples}
\alias{extract_samples}
\title{Helper function to extract samples from chain list,
separates burn_in samples from rest,
and concatenates the chain list to one long chain}
\usage{
extract_samples(chain_list, first_sample)
}
\description{
Helper function to extract samples from chain list,
separates burn_in samples from rest,
and concatenates the chain list to one long chain
}
