% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-MCMC.R
\name{extract_sol_matrix}
\alias{extract_sol_matrix}
\title{Helper function to extract solution matrix from chain list,
separates burn_in samples from rest,
and concatenates the chain list to one long chain}
\usage{
extract_sol_matrix(chain_list, first_sample)
}
\description{
Helper function to extract solution matrix from chain list,
separates burn_in samples from rest,
and concatenates the chain list to one long chain
}
