% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_op.R
\name{forward_op}
\alias{forward_op}
\title{Forward operator in Bayesian inference for linear multibox models}
\usage{
forward_op(X, params, sol = NULL)
}
\description{
Solves the EBM and evaluates it at the time points given in \code{obs_grid}.
}
