% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_noise.R
\name{gen_noise}
\alias{gen_noise}
\title{Generate noise (AR(1) plus white)}
\usage{
gen_noise(n, lambda, weights, n_years, sd_white, sd_ar1)
}
\arguments{
\item{n}{Number of samples.}

\item{lambda}{Vector of lambda values.}

\item{weights}{Vector of weights values.}

\item{n_years}{Number of years (i.e. length of the noise).}

\item{sd_white}{SD of white noise.}

\item{sd_ar1}{SD of AR(1) noise.}
}
\value{
Matrix where each row is a sample,
each column corresponds to a point in time.
}
\description{
Generate noise (AR(1) plus white)
}
\examples{
# only noisy time series with a bit of correlation
as.vector(gen_noise(1, 0.1, 1, 51, 0.1, 0.3))
# fit plus noise
solve_ebm(0.1, 1, 0, 0, 10, c(0, rep(5,50))) + as.vector(gen_noise(1, 0.1, 1, 51, 0.1, 0.3))
}
