% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_noise_from_ebm_fit.R
\name{gen_noise_from_ebm_fit}
\alias{gen_noise_from_ebm_fit}
\title{Generate AR(1) noise for class \code{ebm_fit_result}}
\usage{
gen_noise_from_ebm_fit(n, res_list)
}
\arguments{
\item{n}{Number of samples.}

\item{res_list}{Object of class \code{ebm_fit_result}}
}
\value{
Matrix where each row is a sample of the noise,
each column corresponds to a point in time.
}
\description{
Wrapper for \code{gen_noise}.
Extracts the posterior means from the \code{ebm_fit_result} and uses them in \code{gen_noise}
to generate samples of AR(1) noise. The parameters \code{lambda} and \code{weights}
of the AR(1) noise equal the posterior means.
The SD of the AR(1) noise is chosen s.t. it equals the SD of the residual from the fit.
}
