% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_post_means.R
\name{get_post_means}
\alias{get_post_means}
\title{Helper function to extract posterior means from \code{ebm_fit_result}}
\usage{
get_post_means(res_list)
}
\arguments{
\item{res_list}{Object of class \code{ebm_fit_result}.}
}
\value{
List with entries
\itemize{
\item \code{lambda}: Vector of posterior means for all \code{lambda} parameters
\item \code{weights}: Vector of posterior means for all \code{weights} parameters (including the first one,
}
}
\description{
Helper function to extract posterior means from \code{ebm_fit_result}
}
