% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hadcrut}
\alias{hadcrut}
\title{Temperature observations from Hadcrut5}
\format{
A data frame with 172 rows and 2 variables:
\describe{
\item{year}{from 1850-2022}
\item{temperature}{temperature relative to 1961-1990}
}
}
\source{
\url{https://www.metoffice.gov.uk/hadobs/hadcrut5/}
}
\usage{
hadcrut
}
\description{
A dataset containing the global annual mean surface temperature.
}
\keyword{datasets}
