% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_likelihood.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Log-likelihood function for linear multibox models}
\usage{
log_likelihood(X, params, sol = NULL)
}
\arguments{
\item{X}{Vector of parameter values.}

\item{params}{List of parameters; relevant entries:
\itemize{
\item \code{y_obs}: vector of temperature observations
\item \code{noise_process}: \code{"none"} (then only white noise is considered)
or \code{"fixed_cov"} (then noise where the inverse of the covariance is \code{sigma_e_inv})
\item \code{meas_noise}: (only if \code{noise_process = "none"}) SD of the white noise
\item \code{sigma_e_inv}: (only if \code{noise_process = "fixed_cov"})
Inverse of covariance matrix
\item all parameters relevant for the forward operator \code{forward_op}.
}}

\item{sol}{(optional) Vector of EBM solution for this value of X.
This can be used to speed up the code and avoid repeated calculations.}
}
\value{
Value of the log of likelihood density.
}
\description{
Log-likelihood function for linear multibox models
}
