% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_chain.R
\name{MCMC_chain}
\alias{MCMC_chain}
\title{Compute a single Markov chain with the Metropolis Hastings algorithm}
\usage{
MCMC_chain(X_start, params)
}
\arguments{
\item{X_start}{Initial values for parameters}

\item{params}{List of parameters;
relevant entries (for more details, see vignette for config file):
\itemize{
\item \code{alpha}: weighting factor for adaptive covariance scheme
\item \code{s}: proposal variance
\item \code{M}: number of samples
}}

\item{X_names}{variable names of corresponding parameters}
}
\value{
List with entries:
\itemize{
\item \code{samplesMCMC}: Matrix of samples, one row for each sample, one column for each variable
\item \code{sol_matrix}: Matrix of model solutions, one row for each sample, one column for each year
}
}
\description{
Called by \code{MCMC_run}.
}
