% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_parallel_chains.R
\name{MCMC_parallel_chains}
\alias{MCMC_parallel_chains}
\title{Function to run several chains with dispersed starting points,
possibly in parallel}
\usage{
MCMC_parallel_chains(params)
}
\arguments{
\item{params}{Parameter list. In addition to those relevant in \code{MCMC_chain},
the following entries are required:
\itemize{
\item \code{parallel}: Boolean, should the computation run in parallel_
\item \code{X_names}
\item \code{n_chains}: Number of chains
}}
}
\description{
Function to run several chains with dispersed starting points,
possibly in parallel
}
