% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{my_mm}
\alias{my_mm}
\title{Simple C++ code for matrix-vector multiplication}
\usage{
my_mm(m, v)
}
\arguments{
\item{m}{matrix}

\item{v}{vector}
}
\value{
matrix product (but in a fast way)
}
\description{
Simple C++ code for matrix-vector multiplication
}
