% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_sens_single_it.R
\name{noise_sens_single_it}
\alias{noise_sens_single_it}
\title{Single iteration to check influence of noise sensitivity}
\usage{
noise_sens_single_it(
  forc,
  vars,
  sd_white,
  sd_ar1,
  start_year = NULL,
  end_year = NULL,
  config_file = system.file("extdata/ebm_fit_config.yml", package = "ClimBayes"),
  config = "noise_sens"
)
}
\arguments{
\item{forc}{Vector of forcing values or data frame with columns \code{year} and \code{forcing}.}

\item{vars}{list of data-generating variables with entries
\itemize{
\item \code{lambda}
\item \code{weights}
\item \code{Cap}
\item \code{T0}
\item \code{F0}
}}

\item{sd_white}{SD of white noise}

\item{sd_ar1}{SD of AR(1) noise}

\item{start_year}{First of forcing considered (only relevant if \code{forc} is a data frame).}

\item{end_year}{Last year of forcing considered (inclusive; only relevant if \code{forc} is a data frame).}

\item{config_file}{Path to config file or full config list.}

\item{config}{Name of config to use.}
}
\value{
Object of class \code{ebm_fit} - the result of the fit to the synthetic observations.
}
\description{
Generates synthetic data as the EBM solution with given variables \code{vars} and forcing \code{forc}
plus noise with SDs \code{sd_white} and \code{sd_ar1}. Then fits the EBM to these synthetic observations
and returns the result.
}
\examples{
vars = list(lambda = 0.1, Cap = 10.1, weights = 1, T0 = 0, F0 = 0)
noise_sens_single_it(c(0, numeric(99) + 10), vars, 0.1, 0, 1, 100)
}
