% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_sensitivity.R
\name{noise_sensitivity}
\alias{noise_sensitivity}
\title{Analyze influence of noise in observational data on parameter estimates}
\usage{
noise_sensitivity(
  sd_white_list,
  sd_ar1_list,
  reps,
  forc,
  vars,
  start_year = NULL,
  end_year = NULL,
  config_file = system.file("extdata/ebm_fit_config.yml", package = "ClimBayes"),
  config = "noise_sens"
)
}
\arguments{
\item{sd_white_list}{Vector of SD values for white noise.}

\item{sd_ar1_list}{Vector of SD values for AR(1) noise.}

\item{reps}{Number of repetitions.}

\item{forc}{Vector of forcing values or data frame with columns "year" and "forcing".}

\item{vars}{List of data-generating variables with entries
\itemize{
\item lambda
\item weights
\item Cap
\item T0
\item F0
}}

\item{start_year}{First of forcing considered (only relevant if \code{forc} is a data frame).}

\item{end_year}{Last year of forcing considered (inclusive; only relevant if \code{forc} is a data frame).}

\item{config_file}{Path to config file or full config list.}

\item{config}{Name of config to use.}
}
\value{
Tibble with one row for each iteration and columns:
\itemize{
\item \code{sd_white}: SD of white noise used in data-generation
\item \code{sd_ar1}: SD of AR(1) noise used in data-generation
\item \code{rep}: Number of repetition
\item \code{delta_lambda1}: Difference of data-generating value (i.e. that in \code{vars$lambda[1]})
and the parameter estimate
\item (possibly further columns \code{delta_lambda2} etc.)
}
}
\description{
Function computes uncertainties in Bayesian parameter estimates due to noise in observations.
It generates synthetic data as the EBM solution with given variables \code{vars} and forcing \code{forc}
plus noise with SDs \code{sd_white} and \code{sd_ar1}. Then fits the EBM to these synthetic observations
and calculates the difference of the data-generating parameter values and the estimates.
This is repeated for each combination of SDs in \code{sd_white_list} and \code{sd_ar1_list} and \code{reps} times.
Internally, the function uses \code{noise_sens_single_it}.
}
\examples{
vars = list(lambda = 0.1, Cap = 10.1, weights = 1, T0 = 0, F0 = 0)
noise_sensitivity(c(0, 0.1), c(0, 0.1), 10,
forc = c(0, numeric(99) + 10), vars = vars)
}
