% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ode_helper}
\alias{ode_helper}
\title{Helper function to solve the ODE}
\usage{
ode_helper(n_years, lambda, weights, forc_vals, F0)
}
\arguments{
\item{n_years}{number of years considered}

\item{lambda}{vector for lambda, each entry corresponding to one box}

\item{weights}{vector of weights values, one for each box; should sum up to 1}

\item{forc_vals}{vector of forcing values (yearly data)}

\item{F0}{initial forcing value}
}
\value{
vector of solution values,
not scaled be heat capacity yet
}
\description{
Helper function to solve the ODE
}
