% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_from_config.R
\name{params_from_config}
\alias{params_from_config}
\title{Convert config list into parameter list}
\usage{
params_from_config(n_boxes, n_years, config_list)
}
\arguments{
\item{n_boxes}{Number of boxes of multibox EBM.}

\item{n_years}{Number of years considered.}

\item{config_list}{A nested list of configuration parameters. This should
only be a single config, i.e. it usually is a sublist of
\code{yaml::read_yaml("ebm_fit_config.yml")}.}
}
\value{
List of parameters.
}
\description{
Helper function to extract internal parameter list from config list
}
\examples{
params_from_config(1, 100, yaml::read_yaml("ebm_fit_config.yml")$default)
}
