% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ebm_fit_result.R
\name{plot.ebm_fit_result}
\alias{plot.ebm_fit_result}
\title{Plot function for class \code{ebm_fit_result}.}
\usage{
\method{plot}{ebm_fit_result}(res_list, ...)
}
\arguments{
\item{res_list}{Object of class \code{ebm_fit_result}.}

\item{...}{Additional parameters to be passed on to \code{plot_marginals} or \code{plot_fit}.}
}
\value{
A list with entries
\itemize{
\item \code{gg_fit}: plot of the observations and the model fit
\item \code{gg_marginals}: plot of the marginal posteriors.
}
}
\description{
Plot function for class \code{ebm_fit_result}.
}
