% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ebm_proj_result.R
\name{plot.ebm_proj_result}
\alias{plot.ebm_proj_result}
\title{Plot function for class \code{ebm_proj_result}}
\usage{
\method{plot}{ebm_proj_result}(res_list, line_size = 0.8,
  text_size = 10, cred_int = TRUE, plot_forcing = FALSE)
}
\arguments{
\item{res_list}{Object of class \code{ebm_proj_result} (result of \code{ebm_projection}).}

\item{line_size}{Line size in plot.}

\item{text_size}{Text size in plot.}

\item{cred_int}{Boolean to decide if credible intervals should be plotted.}

\item{plot_forcing}{Boolean to decide if forcing should be plotted below.}
}
\value{
ggplot
}
\description{
Plot function for class \code{ebm_proj_result}
}
