% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ebm_fit_result.R
\name{plot_fit}
\alias{plot_fit}
\title{Function to plot observations and model fit from \code{ebm_fit}.}
\usage{
plot_fit(res_list, cred_int = TRUE, ebm_sol = NULL, text_size = 10,
  line_size = 0.5, plot_forcing = FALSE, ...)
}
\arguments{
\item{res_list}{Object of class \code{ebm_fit_result}.}

\item{cred_int}{Boolean to decide if credible intervals should be plotted.}

\item{ebm_sol}{Vector of additional temperature values,
e.g. the true non-noisy EBM solution in case the observations are created synthetically
as solution + noise (see also the \code{tutorial}-vignette).}

\item{text_size}{Text size in plot.}

\item{line_size}{Line size in plot.}

\item{plot_forcing}{Boolean to decide if forcing should be plotted below.}
}
\value{
A ggplot if \code{plot_forcing = FALSE} and the output of \code{cowplot::plot_grid}
if \code{plot_forcing = TRUE}.
}
\description{
Function to plot observations and model fit from \code{ebm_fit}.
}
