% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_joint_posterior.R
\name{plot_joint_posterior}
\alias{plot_joint_posterior}
\title{Plot function for joint posteriors}
\usage{
plot_joint_posterior(res_list, var1 = "lambda1", var2 = "lambda2",
  xlim = NA, ylim = NA, prior_bounds_as_lim = FALSE,
  text_size = 10, colourbar_breaks = NA)
}
\arguments{
\item{res_list}{Result of fit.}

\item{var1}{Name of first variable, e.g. \code{lambda1}}

\item{var2}{Name of second variable, e.g. \code{lambda2}}

\item{xlim}{Vector for limits of x axis. If \code{NA}, will be determined automatically,
or via \code{prior_bounds_as_lim} (see below).}

\item{ylim}{Vector for limits of y axis}

\item{prior_bounds_as_lim}{If TRUE, the bounds of prior intervals will be taken as
the limits of the plot axes. If FALSE and also no \code{xlim} / \code{ylim} specified,
then the limits are determined automatically.}

\item{text_size}{Text size in plot.}

\item{colourbar_breaks}{Breaks for \code{scale_fill_distiller}.}
}
\description{
Plot of joint posterior of two variables.
}
\examples{
plot_joint_posterior(res_list, "lambda1", "lambda2")
}
