% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_one_vs_twobox.R
\name{plot_one_vs_two_box}
\alias{plot_one_vs_two_box}
\title{Plot function to compare one- with two-box}
\usage{
plot_one_vs_two_box(res_list_1box, res_list_2box, cred_int = TRUE,
  print_RMSE = TRUE, legend_x = 0.15, legend_y = 0.85,
  text_size = 10, line_size = 0.5, cols = c(Observations =
  "darkgrey", `1-box` = "firebrick", `2-box` = "#E95F0A"))
}
\arguments{
\item{res_list_1box}{Result of one-box fit.}

\item{res_list_2box}{Result of two-box fit.}

\item{cred_int}{Boolean to decide if credible intervals should be plotted.}

\item{print_RMSE}{Boolean to decide if root mean square error (RMSE) should be
plotted in a separate text box.}

\item{legend_x}{Legend position on x-axis (passed to the plot via \code{ggplot2::theme(legend.position = c(legend_x, legend_y))}.}

\item{legend_y}{Legend position on y-axis (passed to the plot via \code{ggplot2::theme(legend.position = c(legend_x, legend_y))}..}

\item{text_size}{Text size in plot.}

\item{line_size}{Line size in plot.}

\item{cols}{A named vector to specify colors for \code{Observations}, \code{1-box}, \code{2-box}.}
}
\value{
A ggplot object.
}
\description{
Plot function to compare one- with two-box
}
