% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_two_fits.R
\name{plot_two_fits}
\alias{plot_two_fits}
\title{Plot function to compare two fits to the same observation data.}
\usage{
plot_two_fits(
  res_list1,
  res_list2,
  label1 = "1-box",
  label2 = "2-box",
  cred_int = TRUE,
  print_RMSE = TRUE,
  legend_x = 0.15,
  legend_y = 0.85,
  text_size = 10,
  line_size = 0.5,
  cols = c("darkgrey", "firebrick", "#E95F0A"),
  subtract_mean = FALSE
)
}
\arguments{
\item{res_list1}{Result of first fit.}

\item{res_list2}{Result of second fit.}

\item{label1}{Label for first fit, e.g. \code{"1-box"}.}

\item{label2}{Label for second fit, e.g. \code{"2-box"}.}

\item{cred_int}{Boolean to decide if credible intervals should be plotted.}

\item{print_RMSE}{Boolean to decide if root mean square error (RMSE) should be
plotted in a separate text box.}

\item{legend_x}{Legend position on x-axis (passed to the plot via \code{ggplot2::theme(legend.position = c(legend_x, legend_y))}.}

\item{legend_y}{Legend position on y-axis (passed to the plot via \code{ggplot2::theme(legend.position = c(legend_x, legend_y))}..}

\item{text_size}{Text size in plot.}

\item{line_size}{Line size in plot.}

\item{cols}{A vector to specify colors for observations, the first and the second fit}

\item{subtract_mean}{Boolean to specify if overall mean of all time series should be subtracted before plotting}
}
\value{
A ggplot object.
}
\description{
Plot function to compare two fits to the same observation data.
}
