% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ebm_fit_result.R
\name{print.ebm_fit_result}
\alias{print.ebm_fit_result}
\title{Print method for class \code{ebm_fit_result}.}
\usage{
\method{print}{ebm_fit_result}(res_list)
}
\description{
Print method for class \code{ebm_fit_result}.
}
