% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ebm_proj_result.R
\name{print.ebm_proj_result}
\alias{print.ebm_proj_result}
\title{Print method for class \code{ebm_proj_result}.}
\usage{
\method{print}{ebm_proj_result}(res_list)
}
\description{
Print method for class \code{ebm_proj_result}.
}
