% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{schmidt}
\alias{schmidt}
\title{Forcing data from Schmidt et. al.}
\format{
A data frame with 1151 rows and 2 variables:
\describe{
\item{year}{from 850-2000}
\item{forcing}{radiative forcing in W/m^2 relative to 850}
}
}
\source{
\url{https://doi.org/10.5194/gmd-5-185-2012}
}
\usage{
schmidt
}
\description{
A dataset containing the global radiative forcing anomalies
}
\keyword{datasets}
