% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ebm_wrapper.R
\name{solve_ebm_wrapper}
\alias{solve_ebm_wrapper}
\title{Wrapper for \code{solve_ebm}}
\usage{
solve_ebm_wrapper(X, params)
}
\arguments{
\item{X:}{vector of parameter values.}

\item{params:}{List of parameters.
Relevant parameters:
\itemize{
\item \code{X_names}: parameter names corresponding to the values in \code{X}
(e.g. the default for a one-box model will be \code{X_names = c("lambda1", "F0")},
and \code{X} is of length two)
default values of parameters which are not estimated (e.g.
\code{Cap_default}),
\item \code{forc_vals} (vector of forcing values)
}}
}
\value{
: Vector of EBM solution.
}
\description{
General EBM solver which extracts correct arguments for \code{solve_ebm}
and calls it.
Usually the wrapper is not called by the user,
but instead used in other MCMC functions.
}
